Optimize parallax and tilt animations for performance (#2365)
