+++ apps/gs-web/src/layouts/WebLayout.astro
@@ -26,7 +26,7 @@ const { title = 'Gold Shore', description = '' } = Astro.props;
 
     <GlobalModal />
 
-    <script is:inline>
+    <script>
       import { initGoldShoreUI } from "@goldshore/theme/runtime";
       initGoldShoreUI();
     </script>
