+++ packages/theme/styles/base.css
@@ -1,4 +1,18 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=Playfair+Display:wght@500;600;700&display=swap");
+@font-face {
+  font-family: "Atkinson";
+  src: url("../../../public/fonts/atkinson-regular.woff") format("woff");
+  font-style: normal;
+  font-weight: 400;
+  font-display: swap;
+}
+
+@font-face {
+  font-family: "Atkinson";
+  src: url("../../../public/fonts/atkinson-bold.woff") format("woff");
+  font-style: normal;
+  font-weight: 700;
+  font-display: swap;
+}
 
 *,
 *::before,
@@ -47,13 +61,13 @@ h1 {
   font-size: clamp(2.5rem, 5.8vw, 4rem);
   line-height: 1.02;
   margin: 0 0 1.25rem 0;
-  font-weight: 650;
+  font-weight: 700;
 }
 
 h2 {
   font-size: clamp(1.75rem, 3.4vw, 2.2rem);
   margin: 0 0 0.75rem 0;
-  font-weight: 600;
+  font-weight: 700;
 }
 
 p {
